+++ src/util/image.rs
@@ -6,7 +6,7 @@ use std::{
 
 static OUT_PATH: &str = "out/";
 
-pub fn save_png(image_name: &str, width: u32, height: u32, pixels: Vec<u32>) {
+pub fn save_png(image_name: &str, width: usize, height: usize, pixels: Vec<u32>) {
     let mut out_path = PathBuf::new();
     out_path.push(OUT_PATH);
     out_path.push(format!("{}.png", image_name));
@@ -15,13 +15,13 @@ pub fn save_png(image_name: &str, width: u32, height: u32, pixels: Vec<u32>) {
     let w = BufWriter::new(file);
 
     // TODO: Check these settings.
-    let mut encoder = png::Encoder::new(w, width, height);
+    let mut encoder = png::Encoder::new(w, width as u32, height as u32);
     encoder.set_color(png::ColorType::Rgb);
     encoder.set_depth(png::BitDepth::Eight);
     encoder.set_compression(png::Compression::Best);
     encoder.set_adaptive_filter(png::AdaptiveFilterType::Adaptive);
 
-    let mut image_buffer: Vec<u8> = Vec::with_capacity((width * height * 3) as usize);
+    let mut image_buffer: Vec<u8> = Vec::with_capacity(width * height * 3);
     for pixel in pixels.iter() {
         let r = (pixel >> 16 & 0xFF) as u8;
         let g = (pixel >> 8 & 0xFF) as u8;
